Create R file for Visualization
R Analysis Script Description
The r_analysis.R file contains the R script used to integrate the visualization generated in Python and perform additional analysis on the Netflix dataset. This script was created and executed in an R environment.

Analysis Steps:
Integration of Python Visualization:

The script loads the image file representing the most watched genres visualization generated in Python.
It integrates the visualization into R for further analysis and display.